[Test]:Modify duration calc
Modify duration calculate
Add nanosleep case

+++ perf_test/inc/clock.h
@@ -1,9 +1,9 @@
 /*
  * @Author: Kk-knight liubin999000@163.com
  * @Date: 2023-02-21 21:13:24
- * @LastEditors: Kk-knight liubin999000@163.com
- * @LastEditTime: 2023-02-21 21:30:25
- * @FilePath: /perf_test/clock.h
+ * @LastEditors: liuyanbin liuyanbin@maxieyetech.com
+ * @LastEditTime: 2023-02-22 10:17:58
+ * @FilePath: /performance/perf_test/inc/clock.h
  * @Description: 这是默认设置,请设置`customMade`, 打开koroFileHeader查看配置 进行设置: https://github.com/OBKoro1/koro1FileHeader/wiki/%E9%85%8D%E7%BD%AE
  */
 #pragma once
@@ -21,7 +21,7 @@ public:
     void Stop();
     void PrintInfo();
 private:
-    uint64_t Duration(timespec &start, timespec &end);
+    double Duration(timespec &start, timespec &end);
 
 private:
     timespec real_time_start_;
